+++ dart/lib/src/hub.dart
@@ -15,6 +15,9 @@ import 'tracing.dart';
 /// Configures the scope through the callback.
 typedef ScopeCallback = void Function(Scope);
 
+/// Called when a transaction is finished.
+typedef OnTransactionFinish = FutureOr<void> Function(ISentrySpan transaction);
+
 /// SDK API contract which combines a client and scope management
 class Hub {
   static SentryClient _getClient(SentryOptions options) {
@@ -347,6 +350,7 @@ class Hub {
     bool? waitForChildren,
     Duration? autoFinishAfter,
     bool? trimEnd,
+    OnTransactionFinish? onFinish,
     Map<String, dynamic>? customSamplingContext,
   }) =>
       startTransactionWithContext(
@@ -360,6 +364,7 @@ class Hub {
         waitForChildren: waitForChildren,
         autoFinishAfter: autoFinishAfter,
         trimEnd: trimEnd,
+        onFinish: onFinish,
         customSamplingContext: customSamplingContext,
       );
 
@@ -372,6 +377,7 @@ class Hub {
     bool? waitForChildren,
     Duration? autoFinishAfter,
     bool? trimEnd,
+    OnTransactionFinish? onFinish,
   }) {
     if (!_isEnabled) {
       _options.logger(
@@ -402,6 +408,7 @@ class Hub {
         waitForChildren: waitForChildren ?? false,
         autoFinishAfter: autoFinishAfter,
         trimEnd: trimEnd ?? false,
+        onFinish: onFinish,
       );
       if (bindToScope ?? false) {
         item.scope.span = tracer;