+++ flutter/test/integrations/native_app_start_integration_test.dart
@@ -1,8 +1,8 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
-import 'package:sentry_flutter/src/integrations/mobile_vitals_integration.dart';
-import 'package:sentry_flutter/src/sentry_native_state.dart';
-import 'package:sentry_flutter/src/sentry_native_wrapper.dart';
+import 'package:sentry_flutter/src/integrations/native_app_start_integration.dart';
+import 'package:sentry_flutter/src/sentry_native.dart';
+import 'package:sentry_flutter/src/sentry_native_channel.dart';
 import 'package:sentry/src/sentry_tracer.dart';
 import 'package:flutter/scheduler.dart';
 
@@ -12,7 +12,7 @@ import '../mocks.mocks.dart';
 const fakeDsn = 'https://abc@def.ingest.sentry.io/1234567';
 
 void main() {
-  group('$MobileVitalsIntegration', () {
+  group('$NativeAppStartIntegration', () {
     late Fixture fixture;
 
     setUp(() {
@@ -21,10 +21,10 @@ void main() {
 
     test('native app start measurement added to first transaction', () async {
       fixture.options.autoAppStart = false;
-      fixture.state.appStartEnd = DateTime.fromMillisecondsSinceEpoch(10);
+      fixture.native.appStartEnd = DateTime.fromMillisecondsSinceEpoch(10);
       fixture.wrapper.nativeAppStart = NativeAppStart(0, true);
 
-      fixture.getMobileVitalsIntegration().call(MockHub(), fixture.options);
+      fixture.getNativeAppStartIntegration().call(MockHub(), fixture.options);
 
       final tracer = fixture.createTracer();
       final transaction = SentryTransaction(tracer);
@@ -33,17 +33,17 @@ void main() {
       final enriched = await processor.apply(transaction) as SentryTransaction;
 
       final expected = SentryMeasurement('app_start_cold', 10);
-      expect(enriched.measurements?[0].name, expected.name);
-      expect(enriched.measurements?[0].value, expected.value);
+      expect(enriched.measurements[0].name, expected.name);
+      expect(enriched.measurements[0].value, expected.value);
     });
 
     test('native app start measurement not added to following transactions',
         () async {
       fixture.options.autoAppStart = false;
-      fixture.state.appStartEnd = DateTime.fromMillisecondsSinceEpoch(10);
+      fixture.native.appStartEnd = DateTime.fromMillisecondsSinceEpoch(10);
       fixture.wrapper.nativeAppStart = NativeAppStart(0, true);
 
-      fixture.getMobileVitalsIntegration().call(MockHub(), fixture.options);
+      fixture.getNativeAppStartIntegration().call(MockHub(), fixture.options);
 
       final tracer = fixture.createTracer();
       final transaction = SentryTransaction(tracer);
@@ -53,41 +53,45 @@ void main() {
       var enriched = await processor.apply(transaction) as SentryTransaction;
       var secondEnriched = await processor.apply(enriched) as SentryTransaction;
 
-      expect(secondEnriched.measurements?.length, 1);
+      expect(secondEnriched.measurements.length, 1);
     });
 
     test('measurements appended', () async {
       fixture.options.autoAppStart = false;
-      fixture.state.appStartEnd = DateTime.fromMillisecondsSinceEpoch(10);
+      fixture.native.appStartEnd = DateTime.fromMillisecondsSinceEpoch(10);
       fixture.wrapper.nativeAppStart = NativeAppStart(0, true);
       final measurement = SentryMeasurement.warmAppStart(Duration(seconds: 1));
 
-      fixture.getMobileVitalsIntegration().call(MockHub(), fixture.options);
+      fixture.getNativeAppStartIntegration().call(MockHub(), fixture.options);
 
       final tracer = fixture.createTracer();
-      final transaction =
-          SentryTransaction(tracer).copyWith(measurements: [measurement]);
+      final transaction = SentryTransaction(tracer).copyWith();
+      transaction.measurements.add(measurement);
 
       final processor = fixture.options.eventProcessors.first;
 
       var enriched = await processor.apply(transaction) as SentryTransaction;
       var secondEnriched = await processor.apply(enriched) as SentryTransaction;
 
-      expect(secondEnriched.measurements?.length, 2);
-      expect(secondEnriched.measurements?.contains(measurement), true);
+      expect(secondEnriched.measurements.length, 2);
+      expect(secondEnriched.measurements.contains(measurement), true);
     });
   });
 }
 
 class Fixture {
   final options = SentryFlutterOptions(dsn: fakeDsn);
-  final wrapper = MockNativeWrapper();
-  final state = SentryNativeState();
+  final wrapper = MockNativeChannel();
+  late final native = SentryNative();
 
-  MobileVitalsIntegration getMobileVitalsIntegration() {
-    return MobileVitalsIntegration(
-      wrapper,
-      state,
+  Fixture() {
+    native.setNativeChannel(wrapper);
+    native.reset();
+  }
+
+  NativeAppStartIntegration getNativeAppStartIntegration() {
+    return NativeAppStartIntegration(
+      native,
       () {
         return SchedulerBinding.instance;
       },