+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Feat: Mobile Vitals - Native App Start (#749)
+* Feat: Mobile Vitals - Native Frames (#772)
 
 ## 6.4.0-beta.3